requirements.txt
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
black==22.1.0
bleach==4.1.0
Bottleneck==1.3.2
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
com-net==1.0.0
cryptography==36.0.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.1.0
docker-pycreds==0.4.0
entrypoints==0.4
executing==0.8.2
Flask==2.0.3
fonttools==4.25.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
imagecorruptions==1.1.2
imageio==2.16.0
impacket==0.9.24
importlib-resources==5.4.0
ipykernel==6.9.1
ipython==8.0.1
ipython-genutils==0.2.0
itsdangerous==2.1.0
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyter-server==1.13.5
jupyterlab==3.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
kiwisolver==1.3.2
ldap3==2.9.1
ldapdomaindump==0.9.3
MarkupSafe==2.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
munkres==1.1.4
mypy-extensions==0.4.3
nbclassic==0.3.5
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
notebook==6.4.8
numexpr==2.8.1
numpy==1.21.2
olefile==0.46
opencv-python==4.5.5.62
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pip==21.2.4
platformdirs==2.5.1
prometheus-client==0.13.1
promise==2.3
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pycparser==2.21
pycryptodomex==3.12.0
Pygments==2.11.2
pyOpenSSL==22.0.0
pyparsing==3.0.4
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
sentry-sdk==1.5.6
setuptools==58.0.4
shortuuid==1.0.8
sip==4.19.13
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
stack-data==0.2.0
termcolor==1.1.0
terminado==0.13.1
testpath==0.5.0
threadpoolctl==2.2.0
tifffile==2022.2.9
tomli==2.0.1
torch==1.10.2
torchsummary==1.5.1
torchvision==0.11.3
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing-extensions==3.10.0.2
urllib3==1.26.8
wandb==0.12.10
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.0.3
wheel==0.37.1
yaspin==2.1.0
zipp==3.7.0
